Fix check for associations declared in MappedSuperclasses
Doctrine does not allow toMany associations to be declared on
MappedSuperclasses. The check for this occurs on the child
class, which correctly checks whether the parent is a MappedSuperclass.
However, it does not check that the field in question was actually
declared on that class. This causes an incorrect exception if there is
an inheritance structure like so:

  Root:        Joined/Single Table Inheritance Entity
  Child:       Mapped Superclass
  Grandchild:  Entity

When there is a toMany association on Root, the Grandchild throws an
exception, assuming that the association has come from Child.

This fix updates the check to also include whether the association has
come from the parent MappedSuperclass in question.